new a ChromeJS.js file
