+++ test/src/window.js
@@ -1,6 +1,6 @@
 const localServer = require('../fixtures/server.js')
 const assert      = require('assert')
-const VBot        = require('../../dist').vbot
+const VBot        = require('../../dist')
 const testPath = `${__dirname}/../`
 describe('window', async () => {
   let vbot
